ONM-1347: Use DBAL in Special class model

+++ libs/models/Special.php
@@ -122,31 +122,28 @@ class Special extends Content
     {
         parent::create($data);
 
-        $data['id'] = $this->id;
-
         if (!array_key_exists('pdf_path', $data)) {
             $data['pdf_path'] = '';
         }
 
-        $sql = "INSERT INTO specials "
-             . "(`pk_special`, `subtitle`, `img1`, `pdf_path`)"
-             . " VALUES (?,?,?,?)";
+        try {
+            $rs = getService('dbal_connection')->insert(
+                "specials",
+                [
+                    'pk_special' => (int) $this->id,
+                    'subtitle'   => $data['subtitle'],
+                    'img1'       => (int) $data['img1'],
+                    'pdf_path'   => $data['pdf_path']
+                ]
+            );
 
-        $values = array(
-            $this->id,
-            $data['subtitle'],
-            (int) $data['img1'],
-            $data['pdf_path']
-        );
+            $this->saveItems($data);
 
-        if ($GLOBALS['application']->conn->Execute($sql, $values) === false) {
+            return $this;
+        } catch (\Exception $e) {
+            error_log($e->getMessage());
             return false;
         }
-        $this->saveItems($data);
-
-        $this->read($this->id);
-
-        return $this;
     }
 
     /**
@@ -214,21 +211,24 @@ class Special extends Content
             $data['pdf_path'] = '';
         }
 
-        $sql = "UPDATE specials SET `subtitle`=?, `img1`=?,  `pdf_path`=? WHERE pk_special=?";
-        $values = array(
-            $data['subtitle'],
-            (int) $data['img1'],
-            $data['pdf_path'],
-            intval($data['id']),
-        );
+        try {
+            $rs = getService('dbal_connection')->update(
+                "specials",
+                [
+                    'subtitle'   => $data['subtitle'],
+                    'img1'       => (int) $data['img1'],
+                    'pdf_path'   => $data['pdf_path']
+                ],
+                [ 'pk_special' => (int) $data['id'] ]
+            );
 
-        if ($GLOBALS['application']->conn->Execute($sql, $values) === false) {
+            $this->saveItems($data);
+
+            return true;
+        } catch (\Exception $e) {
+            error_log($e->getMessage());
             return false;
         }
-
-        $this->saveItems($data);
-
-        return true;
     }
 
     /**
@@ -240,19 +240,27 @@ class Special extends Content
      */
     public function remove($id)
     {
+        if ((int) $id <= 0) return false;
+
         parent::remove($id);
 
-        $sql    = 'DELETE FROM specials WHERE pk_special=?';
-        $values = array(intval($id));
+        try {
+            $rs = getService('dbal_connection')->delete(
+                "specials",
+                [ 'pk_special' => (int) $id ]
+            );
 
-        if ($GLOBALS['application']->conn->Execute($sql, $values) === false) {
-            return false;
-        }
+            if (!$rs) {
+                return false;
+            }
 
-        $sql    = 'DELETE FROM special_contents WHERE fk_special=?';
-        $values = array(intval($id));
+            if (!$this->deleteAllContents($id)) {
+                return false;
+            }
 
-        if ($GLOBALS['application']->conn->Execute($sql, $values) === false) {
+            return true;
+        } catch (\Exception $e) {
+            error_log($e->getMessage());
             return false;
         }
 
@@ -268,7 +276,9 @@ class Special extends Content
      */
     public function saveItems($data)
     {
-        $this->deleteAllContents($data['id']);
+        if (array_key_exists('id', $data) && !empty($data['id'])) {
+            $this->deleteAllContents($data['id']);
+        }
 
         if (isset($data['noticias_left'])) {
             $contents = $data['noticias_left'];
@@ -316,19 +326,26 @@ class Special extends Content
         if ($id == null) {
             return $items;
         }
-
-        $sql = 'SELECT * FROM `special_contents` WHERE fk_special=? ORDER BY position ASC';
-        $rs  = $GLOBALS['application']->conn->Execute($sql, array(intval($id)));
-
-        while (!$rs->EOF) {
-            $items[] = array(
-                'fk_content'   => $rs->fields['fk_content'],
-                'name'         => $rs->fields['name'],
-                'position'     => $rs->fields['position'],
-                'visible'      => $rs->fields['visible'],
-                'type_content' => $rs->fields['type_content'],
+        try {
+            $rs = getService('dbal_connection')->fetchAll(
+                'SELECT * FROM `special_contents` WHERE fk_special=? ORDER BY position ASC',
+                [ $id ]
             );
-            $rs->MoveNext();
+
+            foreach ($rs as $row) {
+                $items[] = array(
+                    'fk_content'   => $row['fk_content'],
+                    'name'         => $row['name'],
+                    'position'     => $row['position'],
+                    'visible'      => $row['visible'],
+                    'type_content' => $row['type_content'],
+                );
+            }
+
+            return $items;
+        } catch (\Exception $e) {
+            error_log($e->getMessage());
+            return false;
         }
 
         return $items;
@@ -351,24 +368,24 @@ class Special extends Content
             return false;
         }
 
-        $sql = "INSERT INTO special_contents "
-            . "(`fk_special`, `fk_content`,`position`,`name`,`visible`,`type_content`)"
-            . " VALUES (?,?,?,?,?,?)";
-        $values  = array(
-            $id,
-            $pkContent,
-            $position,
-            $name,
-            1,
-            $typeContent
-        );
+        try {
+            $rs = getService('dbal_connection')->insert(
+                "special_contents",
+                [
+                    'fk_special'   => $id,
+                    'fk_content'   => $pkContent,
+                    'position'     => $position,
+                    'name'         => $name,
+                    'visible'      => 1,
+                    'type_content' => $typeContent
+                ]
+            );
 
-        $rs = $GLOBALS['application']->conn->Execute($sql, $values);
-        if ($rs === false) {
+            return true;
+        } catch (\Exception $e) {
+            error_log($e->getMessage());
             return false;
         }
-
-        return true;
     }
 
     /**
@@ -384,11 +401,16 @@ class Special extends Content
         if (is_null($id)) {
             return false;
         }
-        $sql    = 'DELETE FROM special_contents WHERE fk_content=? AND fk_special=?';
-        $values = array(intval($contentId), intval($id));
-        $rs     = $GLOBALS['application']->conn->Execute($sql, $values);
 
-        if ($rs === false) {
+        $rs = getService('dbal_connection')->delete(
+            "special_contents",
+            [
+                'fk_content' => (int) $contentId,
+                'fk_special' => (int) $id,
+            ]
+        );
+
+        if (!$rs) {
             return false;
         }
 
@@ -407,10 +429,13 @@ class Special extends Content
         if (is_null($id)) {
             return false;
         }
-        $sql = 'DELETE FROM special_contents WHERE fk_special=?';
-        $rs  = $GLOBALS['application']->conn->Execute($sql, array(intval($id)));
 
-        if ($rs === false) {
+        $rs = getService('dbal_connection')->delete(
+            "special_contents",
+            [ 'fk_special' => (int) $id ]
+        );
+
+        if (!$rs) {
             return false;
         }
